+++ AirLib/src/api/RpcLibClientBase.cpp
@@ -208,7 +208,7 @@ CollisionInfo RpcLibClientBase::simGetCollisionInfo(const std::string& vehicle_n
     return pimpl_->client.call("simGetCollisionInfo", vehicle_name).as<RpcLibAdapatorsBase::CollisionInfo>().to();
 }
 
-void RpcLibClientBase::setLedIntensity(int intensity, const std::string& vehicle_name)
+void RpcLibClientBase::setLedIntensity(float intensity, const std::string& vehicle_name)
 {
     pimpl_->client.call("setLedIntensity", intensity, vehicle_name);
 }
